requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.4
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==2.10
Js2Py==0.71
pycparser==2.20
pyjsparser==2.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.5
whitenoise==5.2.0
